+++ src/components/PersistLogin.jsx
@@ -25,8 +25,8 @@ const PersistLogin = () => {
 	}, []);
 
 	useEffect(() => {
-		console.log(`isLoading: ${isLoading}`);
-		console.log(`aT: ${JSON.stringify(auth?.accessToken)}`);
+		// console.log(`isLoading: ${isLoading}`);
+		// console.log(`aT: ${JSON.stringify(auth?.accessToken)}`);
 	}, [isLoading]);
 
 	return (
