+++ src/components/layout/nav/Nav.module.css
@@ -1,26 +1,55 @@
 .nav {
 	display: flex;
-	justify-content: flex-end;
+	justify-content: space-between;
 	align-items: center;
 	width: 100%;
 	height: 3rem;
+	margin: 0 2rem 0 2rem;
 	background-color: rgb(61, 153, 153);
 }
 
-.button {
-	width: 5.5rem;
+.logoutNav {
+	display: flex;
+}
+
+.logout,
+.home,
+.admin {
+	display: flex;
+	align-items: center;
+	justify-content: center;
 	height: 2rem;
-	font-size: 0.9rem;
+	font-size: 0.8rem;
 	font-weight: 700;
-	color: black;
-	background-color: rgb(214, 210, 200);
 	border: none;
 	border-radius: 5px;
 	margin-right: 20px;
+	text-decoration: none;
+	color: black;
+}
+
+.home,
+.logout {
+	width: 5.5rem;
+	color: black;
+	background-color: rgb(214, 210, 200);
 	cursor: pointer;
 }
 
-.button:hover {
+.home {
+	margin-left: 20px;
+	margin-right: 0px;
+}
+
+.admin {
+	display: flex;
+	text-align: center;
+	line-height: 2rem;
+	color: rgb(165, 40, 40);
+}
+
+.home:hover,
+.logout:hover {
 	background-color: floralwhite;
 	color: cadetblue;
 }
