+++ src/components/private_admin/userlist/UserList.module.css
@@ -1,24 +1,3 @@
-.container {
-	min-height: 30rem;
-	width: 35%;
-	max-width: 35%;
-	text-align: center;
-	border: 2px solid rgb(219, 208, 172);
-	border-radius: 1rem;
-}
-
-ol {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	padding: 0;
-	border-top: 3px solid rgb(219, 208, 172);
-}
-
-.title {
-	cursor: pointer;
-}
-
 li {
 	font-size: 1.2rem;
 	font-weight: 700;
