+++ src/pages/private_pages/admin/Admin.module.css
@@ -15,10 +15,38 @@ h1 {
 	justify-content: space-around;
 }
 
-.todo {
+.userListBox {
+	display: flex;
+	flex-direction: column;
 	min-height: 30rem;
-	width: 35%;
-	max-width: 35%;
+	height: 30rem;
+	max-width: 40%;
+	width: 40%;
+	overflow-y: scroll;
+	text-align: center;
+	border: 2px solid rgb(219, 208, 172);
+	border-radius: 1rem;
+}
+
+ol {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	padding: 0;
+	border-top: 3px solid rgb(219, 208, 172);
+}
+
+.title {
+	cursor: pointer;
+}
+
+.todo {
+	display: flex;
+	flex-direction: column;
+	min-height: 30rem;
+	height: 30rem;
+	max-width: 40%;
+	width: 40%;
 	display: flex;
 	border: 1px solid black;
 }
